python/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file='.\requirements.txt' '.\requirements.in'
#
ageqrp==0.1.0
    # via -r .\requirements.in
aiohappyeyeballs==2.4.3
    # via aiohttp
aiohttp==3.10.10
    # via semantic-kernel
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.6.2.post1
    # via
    #   httpx
    #   openai
    #   starlette
asgiref==3.8.1
    # via openapi-core
attrs==24.2.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
azure-core==1.32.0
    # via azure-identity
azure-identity==1.19.0
    # via semantic-kernel
black==24.10.0
    # via -r .\requirements.in
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
chardet==5.2.0
    # via prance
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via
    #   black
    #   nltk
    #   uvicorn
cloudevents==1.11.0
    # via semantic-kernel
colorama==0.4.6
    # via
    #   click
    #   pytest
    #   tqdm
coverage[toml]==7.6.4
    # via pytest-cov
cryptography==43.0.3
    # via
    #   azure-identity
    #   msal
    #   pyjwt
defusedxml==0.7.1
    # via semantic-kernel
deprecated==1.2.15
    # via
    #   opentelemetry-api
    #   opentelemetry-semantic-conventions
deprecation==2.1.0
    # via cloudevents
distro==1.9.0
    # via openai
docopt==0.6.2
    # via -r .\requirements.in
faker==26.1.0
    # via -r .\requirements.in
fastapi==0.115.3
    # via -r .\requirements.in
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via
    #   httpcore
    #   hypercorn
    #   uvicorn
    #   wsproto
h2==4.1.0
    # via hypercorn
hpack==4.0.0
    # via h2
httpcore==1.0.6
    # via httpx
httpx==0.27.2
    # via openai
hypercorn==0.17.3
    # via -r .\requirements.in
hyperframe==6.0.1
    # via h2
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.5.0
    # via opentelemetry-api
iniconfig==2.0.0
    # via pytest
isodate==0.7.2
    # via openapi-core
itsdangerous==2.2.0
    # via -r .\requirements.in
jinja2==3.1.4
    # via
    #   -r .\requirements.in
    #   semantic-kernel
jiter==0.6.1
    # via openai
joblib==1.4.2
    # via nltk
jsonschema==4.23.0
    # via
    #   openapi-core
    #   openapi-schema-validator
    #   openapi-spec-validator
jsonschema-path==0.3.3
    # via openapi-spec-validator
jsonschema-spec==0.2.4
    # via openapi-core
jsonschema-specifications==2023.7.1
    # via
    #   jsonschema
    #   openapi-schema-validator
lazy-object-proxy==1.10.0
    # via openapi-spec-validator
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
more-itertools==10.5.0
    # via openapi-core
msal==1.31.1
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==1.2.0
    # via azure-identity
multidict==6.1.0
    # via
    #   -r .\requirements.in
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via black
nest-asyncio==1.6.0
    # via semantic-kernel
nltk==3.9.1
    # via -r .\requirements.in
numpy==2.2.1
    # via
    #   -r .\requirements.in
    #   semantic-kernel
openai==1.52.2
    # via
    #   -r .\requirements.in
    #   semantic-kernel
openapi-core==0.18.2
    # via semantic-kernel
openapi-schema-validator==0.6.2
    # via
    #   openapi-core
    #   openapi-spec-validator
openapi-spec-validator==0.7.1
    # via openapi-core
opentelemetry-api==1.29.0
    # via
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   semantic-kernel
opentelemetry-sdk==1.29.0
    # via semantic-kernel
opentelemetry-semantic-conventions==0.50b0
    # via opentelemetry-sdk
packaging==24.1
    # via
    #   black
    #   deprecation
    #   prance
    #   pytest
parse==1.20.2
    # via openapi-core
pathable==0.4.3
    # via
    #   jsonschema-path
    #   jsonschema-spec
pathspec==0.12.1
    # via black
platformdirs==4.3.6
    # via black
pluggy==1.5.0
    # via pytest
portalocker==2.10.1
    # via msal-extensions
prance==23.6.21.0
    # via semantic-kernel
priority==2.0.0
    # via hypercorn
propcache==0.2.0
    # via yarl
psycopg[binary,pool]==3.2.3
    # via -r .\requirements.in
psycopg-binary==3.2.3
    # via psycopg
psycopg-pool==3.2.3
    # via psycopg
pybars4==0.9.13
    # via semantic-kernel
pycparser==2.22
    # via cffi
pydantic==2.9.2
    # via
    #   fastapi
    #   openai
    #   pydantic-settings
    #   semantic-kernel
pydantic-core==2.23.4
    # via pydantic
pydantic-settings==2.7.0
    # via semantic-kernel
pyjwt[crypto]==2.10.1
    # via
    #   msal
    #   pyjwt
pymeta3==0.5.1
    # via pybars4
pytest==8.3.3
    # via
    #   -r .\requirements.in
    #   pytest-asyncio
    #   pytest-cov
pytest-asyncio==0.24.0
    # via -r .\requirements.in
pytest-cov==5.0.0
    # via -r .\requirements.in
python-dateutil==2.9.0.post0
    # via faker
python-dotenv==1.0.1
    # via
    #   -r .\requirements.in
    #   pydantic-settings
python-multipart==0.0.15
    # via -r .\requirements.in
pywin32==308
    # via portalocker
pyyaml==6.0.2
    # via
    #   jsonschema-path
    #   jsonschema-spec
referencing==0.30.2
    # via
    #   jsonschema
    #   jsonschema-path
    #   jsonschema-spec
    #   jsonschema-specifications
regex==2023.12.25
    # via
    #   nltk
    #   tiktoken
requests==2.32.3
    # via
    #   azure-core
    #   jsonschema-path
    #   jsonschema-spec
    #   msal
    #   prance
    #   tiktoken
rfc3339-validator==0.1.4
    # via openapi-schema-validator
rpds-py==0.20.0
    # via
    #   jsonschema
    #   referencing
ruamel-yaml==0.18.6
    # via prance
ruamel-yaml-clib==0.2.12
    # via ruamel-yaml
semantic-kernel==1.17.1
    # via -r .\requirements.in
six==1.16.0
    # via
    #   azure-core
    #   prance
    #   python-dateutil
    #   rfc3339-validator
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
starlette==0.41.2
    # via fastapi
tiktoken==0.8.0
    # via -r .\requirements.in
tqdm==4.66.5
    # via
    #   nltk
    #   openai
typing-extensions==4.12.2
    # via
    #   azure-core
    #   azure-identity
    #   fastapi
    #   openai
    #   opentelemetry-sdk
    #   psycopg
    #   psycopg-pool
    #   pydantic
    #   pydantic-core
tzdata==2024.2
    # via psycopg
urllib3==2.2.3
    # via requests
uvicorn==0.34.0
    # via -r .\requirements.in
werkzeug==3.0.6
    # via openapi-core
wrapt==1.17.0
    # via deprecated
wsproto==1.2.0
    # via hypercorn
yarl==1.16.0
    # via aiohttp
zipp==3.21.0
    # via importlib-metadata
